Add files via upload
1. What is the following program doing? Copy the code by typing it into a file named
Q1.py and write your answer in Python comments. Do not worry about non-integer
inputs for this question.
Try writing the answer to this without actually running the program. On the
midterm and final exam, you will not have access to a computer to write or run
code, and you will have to decipher things like this, so it’s worth practicing early.
If you think there is an error in this program, feel free to note that in your comments
as well.
x = input(‘give me an input ’)
y = input(‘give me another input ’)
print(x + (2 * y) + x)
2. What is the following program doing? Copy the code by typing it into a file named
Q2.py and write your answer in Python comments. Do not worry about non-integer
inputs for this question.
x = input(‘give me an input ’)
y = input(‘give me another input ’)
y = int(y)
print(x + str(2 * y) + x)
3. Write a program that asks the user to input an integer. The program first checks if
the user entered an integer, then checks to see if the integer is even or odd. It prints
the word ‘Even’ if even, and the word ‘Odd’ if odd. Name your program Q3.py.
4. Write a program that asks the user to input an integer. The program first checks if
the user entered an integer, then checks to see if the integer is within 10 (including
10 itself) of 100 or 200. If that is the case, the program is going to print ‘Yes’, else it
will print ‘No’.
For example, 90 should print ‘Yes’. 209 should also print ‘Yes’. 189 should print
‘No’.
Name your program Q4.py.
5. A game (it actually gets used sometimes for screening programmers in tech
interviews) called ‘Fizz Buzz’ is the origin for this next question.
Write a program that asks the user to input a positive integer. Check if the user
entered a positive integer. If the number is divisible by 3, your program should print
’Fizz’. If the number is divisible by 5, your program should print ’Buzz’. If the
number is divisible by 3 and 5, your program should print ’Fizz Buzz’. Finally, if none
of these conditions are satisfied, your program should print the number itself.
Note that the program should only print one thing. For example, if the user enters
30, it should only print ‘Fizz Buzz’.
Name your program Q5.py.
6. Ask the user for two integers. First check if both are integers. If both values are in
the range 30 to 40 (inclusive) or if both values are in the range 40 to 50 (inclusive),
print ”Yay” else print ”Nay”.
Name your program Q6.py.
7. “How Old is My Dog in Human Years?” Calculator: Write a program to calculate a
dog’s age in human years.
Ask the user for an age in dog years and calculate that age in human years. Allow for
float values but check the user’s input. You can use the following rules to
approximately convert a medium-sized dog’s age to human years:
• For the first year, one dog year is equal to 15 human years
• For the first 2 years, each dog year is equal to 12 human years
• For the first 3 years, each dog year is equal to 9.3 human years
• For the first 4 years, each dog year is equal to 8 human years
• For the first 5 years, each dog year is equal to 7.2 human years
• After that, each dog year is equal to 7 human years.
Print the result in the following format, substituting for <dog_age> and
<human_age>: "The given dog age <dog_age> is <human_age> in human years."
Round the result to 2 decimal places. Note: If there is a 0 in the hundredths place,
you can drop it, e.g. 24.00 can be displayed as 24.0.
For example:
• If the user enters 2, the program will print: “The given dog age 2 is 24.0 in
human years.”
• If the user enters 3, the program will print: “The given dog age 3 is 27.9 in
human years.”
• If the user enters 4.5, the program will print: “The given dog age 4.5 is 32.4 in
human years.”
• If the user enters 12.1, the program will print: “The given dog age 12.1 is 85.7
in human years.”
Make sure you take into account the possibility of a user inputting a negative age. In
this case the program should print "Age must be a positive number." instead of the
dog's age. If the user doesn’t input a number, print “<age> is not a valid age.”,
substituting for <age>.
Name your program Q7.py.
8. The National Weather Service uses the following formula to calculate windchill:
Twc = 35.74 + 0.6215Ta − 35.75v0.16 + 0.4275Tav0.16
where Twc is the wind chill temperature (feels like), Ta is the air temperature in
degrees F, and v is the wind speed in mph. Note, this only applies for temperatures
below 50F and wind speed above 5 mph.
Write a program that asks the user for the temperature and wind speed and gives
them a wind chill temperature. Round to two decimal places. If the user gives you a
temperature or a wind speed for which the formula is not applicable, give them a
message that politely tells them why it does not make sense to calculate wind chill
(too hot and/or no wind). Allow for input of float values but check the user’s input.
Name your program Q8.py.
9. Download and debug Q9.py. Correct the errors and reupload the file.
10. Download and debug Q10.py. Correct the errors and reupload the file.