Add files via upload
Write a Python program with at least the following functions. We want you to
demonstrate your understanding of code reusability so you should write a few others --
we call these helper functions. For example, you might want to write a function that
calculates and returns all the factors of a given number.
Please name the required functions exactly as they are named below. We are going to
use automated tests for checking them and any incorrect spelling will result in a loss of
points.
isPrime(x) - function that returns whether or not the given number x is prime. This
function returns a boolean.
isComposite(x) - function that returns whether or not the given number x is composite.
This function returns a boolean. We intend for you to reuse code from the prime
number function here.
isPerfect(x) - function that returns whether or not the given number x is perfect. This
function returns a boolean.
isAbundant(x) - function that returns whether or not the given number x is abundant.
This function returns a boolean.
isTriangular(x) - function that returns whether or not a given number x is triangular.
This function returns a boolean.
isPentagonal(x) - function that returns whether or not a given number x is pentagonal.
This function returns a boolean. We intend for you to reuse code from the triangular
number function here.
isHexagonal(x) - function that returns whether or not a given number is hexagonal. This
function returns a boolean.
isNarcissistic(x) - function that returns whether or not a given number is Narcissistic.
This function returns a boolean.
Define and implement these functions, along with a main function, to get a complete
program. We suggest commenting out sections of the main function in order to check
each function that you are writing.