Add files via upload
A Tower Blaster game starts with a main pile of 60 bricks, each numbered from 1 to 60.
Think of the numbers on the bricks as the width of the bricks. The objective is to be the
first player to arrange 10 bricks in your own tower from lowest to highest (from the top
down), because the tower will be unstable otherwise.
The bricks in the main pile are shuffled at the start and both the user and the computer
are dealt 10 bricks from the main pile. As a player receives each brick, they must place
it on top of their current tower in the order it is received. Yes, initially your tower is
likely to be unstable.
After the first 10 bricks are dealt to the user and the computer, there will be 40 bricks
remaining in the main pile. The top brick of the main pile is turned over to begin the
discarded brick pile.
On each player’s turn, the player chooses to pick up the top brick from the discard pile
or to pick up the top brick from the main pile. The top brick from the discard pile is
known. In other words, the discard pile is ‘face up’ and everyone knows how wide the
top brick is. The main pile is ‘face down’. Choosing the top brick from the main pile can
be risky, because the player does not know what the brick is.
Once a player chooses a brick, either from the discard pile or from the main pile, the
player decides where in the tower to put the brick. The tower is always 10 bricks high,
so placing a brick means that an existing brick in the tower is removed and replaced
with the new brick.
If the player takes a brick from the main pile (the one that is ‘face down’), the player can
reject it and place it in the discard pile. This means that nothing in that player’s tower
changes during that turn. If the player takes a brick from the discard pile (the one that is
‘face up’), the player MUST place it into the tower.
The first player to get their 10 bricks in order wins.
If, at any point, all of the cards have been removed from the main pile of bricks, then all
of the cards in the discard pile are shuffled and moved to the main pile. Then the top
card is turned over to start the new discard pile.