Add files via upload
The basic skeleton of a website is an HTML page. This HTML page is a text file with a
certain format. Taking advantage of this fact, one can take an HTML template and
create multiple pages with different values stored. This is exactly what we are going to
do in this HW.
Many websites use these kinds of scripts to mass generate HTML pages from databases.
We will use a sample text file as our ‘database’.
Our goal will be to take a resume in a simplistic text file and convert it into an HTML file
that can be displayed by a web browser.