Add files via upload
This is a simplified version of a common card game, "21". Blackjack is a better-known
variant.
In this game, the dealer deals two "cards" to each player, one hidden, so that only the
player who gets it knows what it is, and one face up, so that everyone can see it.
(Actually, what the other players see is the total of each other player's cards, not the
individual cards.)
There are four players: one human player (the person playing the game) and three
computer players. To play, the players take turns requesting cards, trying to get as close
to 21 as possible, but not going over 21. A player may pass (ask for no more cards). Once
a player has passed, he or she cannot later ask for another card. When all players have
passed, the game ends.
The winner is the player who has come closest to 21 without exceeding it. In the case of
a tie, or if everyone goes over 21, no one wins.
The game is only played once (so it's actually just one "hand").
The "cards" are the numbers 1 through 10 and they are randomly generated, not drawn
from a deck of limited size.