Add files via upload
You will create a Fraction class in Java to represent fractions and to do fraction
arithmetic. We want you to write tests first. Once the tests are written, you can begin
implementing the Fraction class itself.
To get you used to the idea of unit testing, this homework does not require a main
method. You can create one if you find it useful, however, we will not be grading or
even looking at that code. You should be comfortable enough with the accuracy of your
test cases that you do not need to use print statements or a main method to even do ad
hoc “testing” of your Fraction class.