Add files via upload
Traditional Hangman
Hangman is a 2-player word-guessing game generally played by having one player think of a word and the other player trying to guess that word letter by letter. Read more information about the game here.
Similar to Battleship, we want you to program a computer versus user version of the game. The computer will pick a random word from a dictionary that we will provide to you in a file. Note that we want you to use this dictionary file and only this dictionary file. It makes it easier for our testing.
• Represent the word to guess by using a row of underscores, representing each letter of the word (e.g., ‘dog’ would be shown to the user as _ _ _).
• The user guesses one letter at a time. Every correct letter that they guess should be shown in its correct location(s). Every guess that is incorrect will be shown in a list right underneath the word.
For example, below is an example of a reasonable interface. Feel free to make your program do exactly this or make it look even fancier!

Evil Hangman
The evil version of this game basically exploits the computer’s ability to store a large amount of information. In the traditional version of the game, the computer has to stick to the original word as the user guesses. In the evil version, the computer keeps changing the word in order to make the user’s task harder.
The algorithm that drives Evil Hangman is fairly straightforward. The computer begins by maintaining a list of all words of a particular length. Whenever the player guesses, the computer partitions the words into "word families" based on the positions of the guessed letters in the words of a particular length. For example, if the full word list is ECHO, HEAL, BELT, HELLO, and HAIRCUT and the player guesses the letter 'E', then there would be four word families:
E - - -, containing ECHO - E - -, containing HEAL and BELT - E - - -, containing HELLO - - - - - - -, containing HAIRCUT
Once the words are divided into these groups, the computer picks the largest of the groups (with the most words) to use as its remaining word list. By doing so, it gives itself the maximum chance to dodge the user’s guesses.
It then reveals the letters in the positions indicated by the word family. In this case, the computer would pick the family - E - - (because it has 2 words and the other families only have 1 word each) and would reveal an E in the second position of the word.
The computer cannot reverse its decision once it has determined which letters to reveal. In the example above, if the user guesses ‘L’ next, the computer must pick a word from its word list that has four letters and contains an E in the second position. The computer cannot switch and pick the word HELLO. Since both HEAL and BELT have
an L, the computer must pick one of these two words and reveal the L in the correct position.